chore(ci): fix generic display test guard fallout
